Backend/requirements.txt
Django==4.2.20
djangorestframework==3.15.1
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.1
drf-spectacular==0.26.3
pillow==11.1.0
django-cors-headers==4.7.0
